Extract to yamlschema
